pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fleet-management"
version = "0.1.0"
description = "Fleet management optimization with Gaussian degradation via MILP"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "gurobipy",
    "pyyaml",
    "h5py",
    "matplotlib",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["src"]
